Use module to avoid monkey patch
